fix: Make datalist for color-input more understandable
The "datalist" input did not make any sense to users. Made a custom solution for color inputs, where I explain what inputs to add, and that the list is optional.
(The datalist is a way to specify which colors the user can select in a color picker. When absent, the browser defaults are used.)