feat: Streamline the form configuration.
Split the config-form into "Basic", "Submit-settings" and "Advanced settings", and kept only the most important configuration in the top of the form.
The goal is that the user only has to use the first field-set for trivial cases, use the second for the common cases, and only use the last field-set in the most advanced cases.
Removed the column feature, as it did not work properly. Also, multi-columns makes for bad usability, and should be avoided.
Also fixed some problems with setting submit-button text. (Incorrect use of thymeleaf)